requirements.txt
absl-py==0.12.0
astor==0.8.1
astunparse==1.6.3
awscli==1.19.44
botocore==1.20.44
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
cycler==0.10.0
docopt==0.6.2
docutils==0.15.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.10.0
imutils==0.5.4
jmespath==0.10.0
joblib==1.0.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
MplayerCtrl==0.3.3
mss==6.1.0
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.1.0
pipreqs==0.4.10
protobuf==3.15.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-vlc==3.0.11115
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.6
scikit-learn==0.21.2
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
tensorboard==1.14.0
tensorboard-plugin-wit==1.8.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==2.3.1
termcolor==1.1.0
tqdm==4.43.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
wxPython==4.1.1
yarg==0.1.9
zipp==3.4.1
